add dependencies and fix url and methods

+++ main.go
@@ -4,6 +4,7 @@ import (
     "fmt"
     "html/template"
     "net/http"
+    "strconv"
 
 	"app/db"
 )
@@ -11,9 +12,8 @@ import (
 func main() {
     db.InitializeDB()
     http.HandleFunc("/units", unitsHandler)
-    http.HandleFunc("/units/create", createUnitHandler)
-    http.HandleFunc("/units/{id}", detailUnitHandler)
-    http.HandleFunc("/units/{id}/delete", deleteUnitHandler)
+    http.HandleFunc("/units/{id}", unitHandler)
+    http.HandleFunc("/units/{id}/add_dependency", addDependencyHandler)
     
     http.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
         http.ServeFile(w, r, "web/index.html")
@@ -24,32 +24,28 @@ func main() {
 }
 
 func unitsHandler(w http.ResponseWriter, r *http.Request) {
-    // Parsea la plantilla base y la específica
-    tmpl, err := template.ParseFiles("web/base.html", "web/units.html")
-    if err != nil {
-        http.Error(w, err.Error(), http.StatusInternalServerError)
-        return
-    }
-
-    // Datos que se pasarán a la plantilla
-    data := struct {
-        Title string
-        Units []db.Unit
-    }{
-        Title: "Página de Unidades",
-        Units: db.GetUnits(),
-    }
-
-    // Ejecuta la plantilla, automáticamente "extiende" la base
-    err = tmpl.ExecuteTemplate(w, "base.html", data)
-    if err != nil {
-        http.Error(w, err.Error(), http.StatusInternalServerError)
-    }
-}
-
-func createUnitHandler(w http.ResponseWriter, r *http.Request) {
     if r.Method == "GET" {
-        unitsHandler(w, r)
+        // Parsea la plantilla base y la específica
+        tmpl, err := template.ParseFiles("web/base.html", "web/units.html")
+        if err != nil {
+            http.Error(w, err.Error(), http.StatusInternalServerError)
+            return
+        }
+
+        // Datos que se pasarán a la plantilla
+        data := struct {
+            Title string
+            Units []db.Unit
+        }{
+            Title: "Página de Unidades",
+            Units: db.GetUnits(),
+        }
+
+        // Ejecuta la plantilla, automáticamente "extiende" la base
+        err = tmpl.ExecuteTemplate(w, "base.html", data)
+        if err != nil {
+            http.Error(w, err.Error(), http.StatusInternalServerError)
+        }
     } else if r.Method == "POST" {
         r.ParseForm()
         name := r.Form.Get("name")
@@ -66,40 +62,58 @@ func createUnitHandler(w http.ResponseWriter, r *http.Request) {
     }
 }
 
-func detailUnitHandler(w http.ResponseWriter, r *http.Request) {
-    // Parsea la plantilla base y la específica
-    tmpl, err := template.ParseFiles("web/base.html", "web/unit.html")
-    if err != nil {
-        http.Error(w, err.Error(), http.StatusInternalServerError)
-        return
-    }
 
-    // Obtiene el ID de la URL
-    id := 0
-    fmt.Sscanf(r.URL.Path, "/units/%d", &id)
+func unitHandler(w http.ResponseWriter, r *http.Request) {
+    if r.Method == "GET" {
+        // Parsea la plantilla base y la específica
+        tmpl, err := template.ParseFiles("web/base.html", "web/unit.html")
+        if err != nil {
+            http.Error(w, err.Error(), http.StatusInternalServerError)
+            return
+        }
 
-    // Datos que se pasarán a la plantilla
-    data := struct {
-        Title string
-        Unit  db.Unit
-    }{
-        Title: "Detalle de la Unidad",
-        Unit:  db.GetUnit(id),
-    }
+        // Obtiene el ID de la URL
+        id := 0
+        fmt.Sscanf(r.URL.Path, "/units/%d", &id)
 
-    // Ejecuta la plantilla, automáticamente "extiende" la base
-    err = tmpl.ExecuteTemplate(w, "base.html", data)
-    if err != nil {
-        http.Error(w, err.Error(), http.StatusInternalServerError)
+        // Datos que se pasarán a la plantilla
+        data := struct {
+            Title string
+            Unit  db.Unit
+            Dependencies []db.Unit
+            Units []db.Unit
+        }{
+            Title: "Detalle de la Unidad",
+            Unit:  db.GetUnit(id),
+            Dependencies: db.GetDependencies(id),
+            Units: db.GetUnits(),
+        }
+
+        // Ejecuta la plantilla, automáticamente "extiende" la base
+        err = tmpl.ExecuteTemplate(w, "base.html", data)
+        if err != nil {
+            http.Error(w, err.Error(), http.StatusInternalServerError)
+        }
+    } else if r.Method == "DELETE" {
+        id := 0
+        fmt.Sscanf(r.URL.Path, "/units/%d", &id)
+
+        db.DeleteUnit(id)
+
+        http.Redirect(w, r, "/units", http.StatusSeeOther)
     }
 }
 
-func deleteUnitHandler(w http.ResponseWriter, r *http.Request) {
-    // Obtiene el ID de la URL
-    id := 0
-    fmt.Sscanf(r.URL.Path, "/units/%d/delete", &id)
+func addDependencyHandler(w http.ResponseWriter, r *http.Request) {
+    if r.Method == "PUT" {
+        id := 0
+        fmt.Sscanf(r.URL.Path, "/units/%d/add_dependency", &id)
 
-    db.DeleteUnit(id)
+        r.ParseForm()
+        depends_on_id, _ := strconv.Atoi(r.Form.Get("depends_on_id"))
 
-    http.Redirect(w, r, "/units", http.StatusSeeOther)
+        db.CreateDependency(id, depends_on_id)
+
+        http.Redirect(w, r, fmt.Sprintf("/units/%d", id), http.StatusSeeOther)
+    }
 }
